CR 3024992: RT3885 OpenSSL fails to cross-compile on 32-bit->64-bit
On alsi6:64 the version of perl cannot handle 64-bit numbers (might be
a 32-bit executable?), so when the x86_64-xlate.pl script encounters
a 64-bit number, the oct() function munges it into a double-precision
rather than a 64-bit integer. In this case, we know it's either a hex
number or an octal number. So, if a 64-bit hex, just pass through as
hex string, otherwise allow oct() to handle it (which are more likely
to be mul/div/mod).

This is a problem in ghash-x86_64.pl (ghash-x86_64.s), and the solution
has an impact all 64-bit platforms. Ran the unit tests to make sure it's
OK

Original Author: tshort@akamai.com
Reviewer: rsalz@akamai.com
Akamai Status: published RT#3885