CR 3278531: Move preferred ciphers to EX_DATA
Remove cipher-based additions to the SSL and SSL_CTX structures, move
them to an EX_DATA-based data structure.
Move APIs and code to ssl_akamai.[ch]

Originally CR 38059, CR 3012455, and CR 3043841

These changes have been offered to upstream, but we are moving them to
our own sets of files to keep the OpenSSL code reasonably clean. If or
when this code is incorporated into upstream, this code will be
updated and/or removed.

This can be rebased with commit 5159a84 for preferred ciphers

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: proprietary