CR 3024992: OpenSSL fails to compile on alsi6:64
The original patch was causing a problem on some x86_64 architechtures
due to mangling of constants generated via this script. Update the
original patch to only process 64-bit numbers, and keep the original
processing for 32-bit numbers (which are more likely to be mul/div/mod)

Original Author: Todd Short (tshort@akamai.com)
Reviewer: rsalz@akamai.com
Akamai Status: To be published

Can be combined with commit 5ab0ba4 on a rebase.