CR 3130351: Export ciphers are broken SSLv2/SSLv3
Some minor logic errors in the asynchronous event processing code
prevented export ciphers from working in SSLv2 and SSLv3.

s3_srvr.c: s->s3->tmp.use_rsa_tmp = 0 should only be done during the
SSL_ST_SW_KEY_EXCH_A state.

s2_srvr.c: the key_length variable, which is supposed to be the full
length of the key, was being overwritten by the result of the RSA
decryption. This is fine for non-export ciphers, but with export
ciphers, the encrypted key length is usually only 40-bits, and the
rest of the key is in the clear. So, by setting key_length to be the
decryption length, the wrong amount of data was being copied later on.
Instead, leave the calculated key_length as is, and do the encrypted/
decrypted length comparison properly.

Should be combined with commit 9d73ec1 when a rebase occurs.

Original Author: tshort@akamai.com
Reviewer: rsalz@akamai.com
Akamai Status: to be published