CR 3152971: Akamai chmods have shell syntax error
Remove the forced copy of the Makefile to Makefile.save.
The force copy of the makefile is due to the default behavior of
Perforce marking the files as read-only to force you into using the
'p4 edit' command. Files under Perforce can be configured to be
writeable upon sync.

Can be merged with commit a57f3d1 upon rebase.

Reviewer: bkaduk@akamai.com
Akamai Status: internal-only