CR 3152971: Akamai chmods have shell syntax error
Do not chmod the Makefiles.
Do not create or chmod opensslconf.h (which is no longer checked in)
Most (if not all) of the chmod additions to the makefiles are due to the
default behavior of Perforce marking the files as read-only to force you
into using the 'p4 edit' command. Files under Perforce can be configured
to be writeable upon sync.

Can be merged with commit 02c0cc63 upon rebase.

Reviewer: bkaduk@akamai.com
Akamai Status: internal-only