CR 3163801: Restore HIGH ciphers
Ghost had a backwards compatibility issue with TLSv1.2 ciphers being
added to the HIGH list; it inadvertently impacted their customers'
configurations. So, the HIGH cipher list was reverted to not include
TLSv1.2 ciphers. This is a drastic change in behavior for OpenSSL.
This was done in CR 1865260 and CR 2361612.

This change restores the HIGH ciphers to OpenSSL's value and adds
an DEP-COMPS option for ghost to enable the behavior it wants:

DEP-COMPS := openssl/enable/ghost-high

This commit can be combined with commit 1dd7834 upon rebase (which may
need to move later, after the componentization changes).

Reviewer: bkaduk@akamai.com
Akamai Status: internal-only