CR 3171951: libopenssl.a is empty when no TLSv1
libopenssl.a is empty when TLSv1 is not included. This is because the
libssl.a library is not made due to the target being build_crypto.
When this happens, libssl.a (along with libopenssl.a) is made in
akamake via "ar qc".
Instead, pull the creation out of Makefile.org, and into akamake
This cleans up Akamai's changes to OpenSSL, and keeps it in akamake

This can be combined with commit 02c0cc6 or commit df8033c (which it
basically undoes) in a rebase.

Code Reviewer: bkaduk@akamai.com
Akamai Status: internal-only