CR 3182951: sha1 always used in ocsp app
The patch for multiple certificates in the ocsp application always
used SHA1, but not all certificates are signed with SHA1. The original
OpenSSL code used 'cert_id_md' for the message digest, so, put that
in there instead of EVP_sha1().

This can be combined with commit a47dbc9

Original Author: chzhang@akamai.com
Code Reviewer: bkaduk@akamai.com