CR 3238341: Don't hash ServerRandom for resumption
The cryptoserver (akamai-async-rsalg) protocol specifies that the
ServerRandom sent to the client is computed as the hash of some
actual random data (modulo timestamps), so that the TLS terminator
can prove freshness to the cryptoserver by providing the pre-hash
value along with the client's submission.  Assuming that in normal
operation, the initial random value is discarded after the session
is established, compromise of the TLS terminator will not allow
disclosure of the master secret for previous sessions, because the
attacker will not know the pre-hash random value and the cryptoserver
will reject such replays.

However, for session resumption, the TLS terminator already has
the session master secret and the cryptoserver will not be involved
in the abbreviated handshake.  In this case, the ServerRandom should
not be hashed, so that the client and TLS terminator use the same
ServerRandom value when computing transport keys.

Check against the (SSL *)->hit variable to detect resumption and
skip the RSALG_hash for resumption.

Original Author: Samuel Erb <serb@akamai.com>
Code Reviewer: tshort@akamai.com
Akamai Status: internal-only