CR 3235831: fix typo in mk1mf.pl
The make rule for generating bn_prime.h erroneously specified
a call to the 'peral' utility, which does not exist.  It was
intending to use perl, which is properly spelled as \$(PERL)
in this context.

Can be combined with commit 6d610bd64 on rebase.

Original Author: Benjamin Kaduk <bkaduk@akamai.com>
Code Reviewer: rsalz@akamai.com
Akamai Status: unpublished