CR MANY: Add SSL/SSL_CTX_use_cert_and_key()
Upstream PR #1130

CR 3264211: Allow clients of CS and ISSM to load cert/public key
CR 3289291: Additional fixes for SSL_use_PrivateKey
CR 3304391: Replace use_no_PrivateKey with new API
CR 3278531: Improve ABI compatibility, use EX_DATA
CR 3332671: Build issues with clang

Add functions that will do the work of assigning certificate, privatekey
and extra certs to an SSL or SSL_CTX. If no privatekey is given, use the
publickey. This will permit the keys to pass validation for both ECDSA
and RSA. If a private key has already been set for the certificate, it
is discarded. A real private key can be set later.

This is an all-or-nothing setting of these parameters. Unlike the
SSL/SSL_CTX_use_certificate() and SSL/SSL_CTX_use_PrivateKey() functions,
the existing cert or privatekey is not modified (i.e. parameters copied).
This permits the existing cert/privatekey to be replaced.

The API is SSL_use_cert_and_key()/SSL_CTX_use_cert_and_key() and is
defined in the .pod file and the CR.
It replaces the sequence of:
* SSL_use_certificate()
* SSL_use_privatekey()
* SSL_set1_chain()
And may actually be faster, as multiple checks are consolidated.

To properly make use of this API, an ENGINE module (or equivalent, e.g.
ISSM or CryptoServer) needs to contain the actual private key that is to
be used.

Note that ECDH (using the certificate's ECDSA key) ciphers do not work
without the private key being present, based on how the private key is
used in ECDH. It is considered significantly risky to put in that change
as well. ECDH does not offer PFS; ECDHE ciphers should be used instead.

In the master branch this caused crashes in unit tests. The crashes
weren't happeing in 1.0.2, but better to be safe than sorry.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: PR #1130

(cherry picked from commit a6e411374befe26ba02ba35201b3a5c5f8948124)

Conflicts:
	include/openssl/ssl.h
	ssl/ssl_err.c