CR 3264211: Add SSL/SSL_CTX_use_no_PrivateKey()
Iterate through the certificates without a private key already set, and
set the private key to be the public key of the certificate. This will
permit the keys to pass validation for both ECDSA and RSA.
If a private key has already been set for the certificate, it is
discarded. A real private key can be set later.

To properly make use of this API, an ENGINE module (or equivalent, e.g.
ISSM or CryptoServer) needs to contain the actual private key that is to
be used.

Note that ECDH (using the certificate's ECDSA key) ciphers do not work
without the private key being present, based on how the private key is
used in ECDH. It is considered significantly risky to put in that change
as well. ECDH does not offer PFS; ECDHE ciphers should be used instead.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: Internal Only