CR 3264211 3289291 3304391: Add SSL/SSL_CTX_use_cert_and_key()
Add functions that will do the work of assigning certificate, privatekey
and extra certs to an SSL or SSL_CTX. If no privatekey is given, use the
publickey. This will permit the keys to pass validation for both ECDSA
and RSA. If a private key has already been set for the certificate, it
is discarded. A real private key can be set later.

This is an all-or-nothing setting of these parameters. Unlike the
SSL/SSL_CTX_use_certificate() and SSL/SSL_CTX_use_PrivateKey() functions,
the existing cert or privatekey is not modified (i.e. parameters copied).
This permits the existing cert/privatekey to be replaced.

The API is SSL_use_cert_and_key()/SSL_CTX_use_cert_and_key() and is
defined in the .pod file and the CR.
It replaces the sequence of:
* SSL_use_certificate()
* SSL_use_privatekey()
* SSL_set1_chain()
And may actually be faster, as multiple checks are consolidated.

To properly make use of this API, an ENGINE module (or equivalent, e.g.
ISSM or CryptoServer) needs to contain the actual private key that is to
be used.

Note that ECDH (using the certificate's ECDSA key) ciphers do not work
without the private key being present, based on how the private key is
used in ECDH. It is considered significantly risky to put in that change
as well. ECDH does not offer PFS; ECDHE ciphers should be used instead.

The missing_parameters functions for dh, dsa, and ec do not check for
NULL.

In the master branch this caused crashes in unit tests. The crashes
weren't happeing in 1.0.2, but better to be safe than sorry.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: Internal Only

CR 3332671: clang: use_cert_and_key API

Fix clang issues SSL_use_cert_and_key/SSL_CTX_use_cert_and_key APIs.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: internal only