CR 3264211: Add SSL/SSL_CTX_use_no_PrivateKey()
Iterate through the certificates without a private key already set, and
set the private key to be the public key of the certificate. This will
permit the keys to pass validation for both ECDSA and RSA.
If a private key has already been set for the certificate, it is
discarded. A real private key can be set later.

To properly make use of this API, an ENGINE module (or equivalent, e.g.
ISSM or CryptoServer) needs to contain the actual private key that is to
be used.
Support for ENGINES is done by replacing a call to RSA_private_decrypt()
with EVP_PKEY_decrypt(). This is needed in 1.0.2 as it's a direct call
to the RSA code, but this behavior changes in 1.1 (master), so it can be
undone in that branch.
To make ECDSA ENGINESs easier to create, corresponding get functions are
defined for ECDSA_METHOD objects.

This change also removes the CrypoServer code that skips the private key
validation (in commit c7356d6), since validation passes with this method.

Note that ECDH (using the certificate's ECDSA key) ciphers do not work
without the private key being present, based on how the private key is
used in ECDH. It is considered significantly risky to put in that change
as well. ECDH does not offer PFS; ECDHE ciphers should be used instead.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: Internal Only