CR 3266751: Fix copy/paste double-free
Due to a copy/paste error, sh.bittable was being freed when
sh.bitmalloc was non-null, potentially for a second time.
This fixes the free statement to match the conditional.
It may be safe to just make the OPENSSL_free() calls unconditional,
but that is a larger change than needed for this patch.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: published on master (which never had the typo)