Revert "CR 3264211: Add SSL/SSL_CTX_use_no_PrivateKey()"
This reverts commit fd0d88599d6f1884eb42c6e22d5e14303e250eb0.

To keep perforce and git in sync, we can't just cut a branch early. We
have to revert this change due to commit d7c3985, and then re-apply
it. Good thing is that this change actually has a bug, so it will be
fixed when it is recommitted.

(If we didn't care, we'd cut the branch after commit bc50257 and
cherry-pick d7c3985 over to that branch.)

Also, it helps to have this change to do the actual revert in
perforce.

On a rebase, this commit and fd0d885 can be discarded in favor of the
new one to be done later.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: Internal only