CR 3171821: (Re-)disable crypto/store by default
In upstream, it is disabled as "experimental" (and indeed, the
implementation is incomplete, leaks memory, etc.)  I am not entirely
sure why it was enabled for Akamai, but it does not seem to be
used by ghost, so it is probably safe to disable it (again).

In case it is actually needed, provide an openssl/enable/store
override.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: Internal only