CR 3264401: Source files deleted on dclean
The original fix was not complete; this impoved upon it. Dependencies
were not be calculated correctly, lint and tags would not be done
correctly, either.
Added a $DSRC variable which contains a list of all the non-local source
that might be symlinked to dummytest.c, and on dclean, subsequently
deleted.
Updated $SRC to be $DSRC + other files that reside in the test directory
Now, the generate and local_depend targets depend on $DSRC rather than
$SRC

Can be combined with commit 59d2c6f on rebase.

Original Author: tshort@akamai.com
Code reviewer: bkaduk@akamai.com
Akamai Status: to be published (if applicable)