CR 3278531: Move cipher method to ssl_akamai
Remove cipher-based additions to the SSL and SSL_CTX structures, move
them to an EX_DATA-based data structure.
Move APIs and code to ssl_akamai.[ch]

These changes have been offered to upstream, but we are moving them to
our own sets of files to keep the OpenSSL code reasonably clean. If or
when this code is incorporated into upstream, this code will be
updated and/or removed.

This can be rebased with commit c927387

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: proprietary