CR 3293371: TLS1.1 and NO RENEG flags use same bit
The SSL_OP_NO_TLSv1_1 and SSL_OP_DISALLOW_RENEGOTIATION flags are
using the same bit. When code was moved from 6.101 to 6.102, the
options field was restored to a long (from a long long), but the
SSL_OP_DISALLOW_RENEGOTIATION flag was not adjusted correctly.

Move the flag to be 0x00200000, which is unused (and was probably
SSL_OP_EPHEMERAL_RSA originally).

During the ABI updates, this flag and the RSALG flag will be moved
to an Akamai API and will no longer be part of the OpenSSL OP space.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: publication needs update