CR 3266321: add PEM_read_bio_flags
Refactor PEM_read_bio to use BIO_mems instead of BUFs directly,
use some helper routines to reduce the overall function length, and make
some of the checks more reasonable.

The added flags field permits the use of secure memory for the temporary storage,
though the code to acutally do so will be added in a follow-up commit.