CR 3320981: NULL-check both EVP_PKEY_cmp_parameters args
Check both arguments for NULL before calling, instead of just the first
one.  As it happens, at the time of this writing, the second one cannot
actually be NULL, but the logic enforcing that is rather fragile (and
requires separate analysis for the kEDH, kDHr, and kDHd cases).  Comply
to best practices and check everything for future-proofing.

Also update the style while we're here.

Coverity ID 14142

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal-only (this code does not exist on master)