CR 3331791: support debug builds on osx
When akamake detects that the 'debug' target is passed on the command
line, we enter a special mode in akamake-env.mak and force the openssl
configuration target to akamai-debug.  This works for linux systems,
but on OS X we need to use a different perlasm scheme (and a different
syntax for selecting 32-bit or 64-bit).  Add some logic to detect
when BUILDENV.OS is osx, and use KERNEL_BITS to control whether
a 32-bit or 64-bit OS X build is to be performed.

While here, fix up the definitions of akamai-debug and akamai-ccmalloc to
use the no_asm_filler placeholder, and have the correct number of fields.
The easiest way to see what the fields are is in Configure, starting with

my $idx = 0;

The indices for the fields are allocated sequentially thereafter, with
34 fields in total (33 colons as separators), so when $no_asm_filler
(which has 18 separators) is used, the specification should include
15 colons.

Original Author: bkaduk@akamai.com
Code Reviewer: pbohman@akamai.com
Akamai Status: internal only