CR 3374441: Update Configuration/akamake for shared
Combination of most of the akamake-related changes for shared libraries.

* Split CONFIGFLAGS into features and flags

We will want to use a specific set of features for shared-library builds,
so use a separate varaible to facilitate reuse of the logic for the other
arguments to Configure, including the sysname and certain CFLAGS.

* Build a shared library by default, but require confirmation

Force a fixed featureset for the shared library build. The fixed
featureset differs between 32- and 64-bit builds, because the
ec_nistp_64_gcc_128 code relies on 128-bit integers, which are only
available in the 64-bit target.

Detect the presence of openssl/enable/feature components for features not
enabled in the shared build, and error out the build with a message
noting that a static build is required for those features.

Our design discussions have gone back and forth a lot between
the conflicting goals of making the upgrade to shared be seamless
and drop-in for consuming components, and of having a forced
interaction to notify component owners/release leads that there
is a significant change requiring process changes.  For now,
we've settled on having the default behavior be to build shared
libraries, but to require the master component to explicitly set
$P/BUILDENV.linux.OPENSSL_LINKAGE-master := shared
There is a long error message printed in the absence of that
setting, with directions on how to proceed, so hopefully the
update will be relatively painless.

Additionally, we are working to put checks into release processes
to help ensure that the needed shared library components are
present in releases with components that link against openssl shared
libraries, even if there is a communication break between the component
owner that upgraded to shared openssl and the release engineer.

* Use BUILDENV semantics to control shared vs. static linkage

Whether the openssl linkage is shared or static is a property of
the build environment, and must be global for all compilations within
a given tree.  In particular, we do not want a single component that
opts out of shared linkage for itself to cause a cascade of static
linkage for all components that include it -- only the master component
should be able to control whether that build is done with shared or
static openssl, which is exactly what the BUILDENV feature of Akamake
is designed for.  Unfortunately, changes to Akamake are required in
order to add new variables to the BUILDENV framework, but it is possible
to get a close approximation without Akamake support, at the cost of
more logic in the openssl makefile.  We can use that logic until our
patches to add support to Akamake for the OPENSSL_LINKAGE BUILDENV
variable are deployed.

In order to do as much as possible on the first (*) pass through akamake,
conditionalize the presented build targets and exported libraries based on
either (1) BUILDENV.OPENSSL_LINKAGE-master, if defined by Akamake, or
(2) whether the master component DEP-COMPS on openssl/enable/static.
We will document that components must set both, and once Akamake has support
for the OPENSSL_LINKAGE buildenv, that will take precedence.  Since we want
the master component to have control over the static opt-out, it is sufficient
to have our build logic conditionalize on the master component's DEP-COMPS.
We will also need to have a check and error out if openssl/enable/static
is present but *not* set by the master component, but that can be done
elsewhere (and must use delayed evaulation).

Provide assertions for those conditions that we expect to only
hold simultaneously.

(*) first real execution, i.e., after akamaster has processed buildenv and
akamake-env.mak and such.

* Export shared libs as a tree

We already export them as EXP-LIBS, but we also need to export them
as an EXP-TREE, so that the openssl-shared component(s) can import
them and put them in a bundle.

Gate it behind a OPENSSL-WANT-EXP-SHARED variable, to be set as
part of the build configuration.

The unversioned path is just a symlink, so it is better to export
the actual files.

* Record libcrypto's dependency on pthreads

As part of our Akamai customizations to libcrypto, we provide a
(builtin) locking callback, that uses pthread mutex operations
and compiler-intrinsic atomic operations.  This causes libcrypto
to require symbols from libpthread, which is not the case in stock
upstream OpenSSL (1.0.x; 1.1.x is a different story).  For static
library builds, when an application links against the static library,
only the object files within the archive that are needed to resolve
symbols that are used by the application are included.  Thus,
so long as none of the locking or thread-support routines are used
in the application, with static linking, no ill effects are
observed.  When the application is linking libcrypto dymanically,
though, all symbols are resolved at link time, without the
"link-time optimization" of pruning unused object files; this
causes the missing dependency to always be visible and to cause
all consumers of libcrypto.so to fail to link (unless they
pull in -lpthread for some other reason).

Properly record the library dependency by calling declare-lib-deps
from akamake and using Configurations.akamai to add -lpthread to
the EX_LIBS make variable.  This is something of a compromise, since
even the default configurations in Configurations are potentially
affected by the missing dependency, unless OPENSSL_NO_AKAMAI is set.

* Add LD_LIBRARY_PATH for tests

Appending to LD_LIBRARY_PATH in the simple way leaves an empty
component if LD_LIBRARY_PATH was unset prior to the append operation;
this causes the current directory to be searched first, which
is usually not desired.

* Update componentinfo.xml to the 7.x version

* Use buildenv-like variables for shared or static

Use
$P/BUILDENV.linux.OPENSSL_LINKAGE-master := shared
for static compilation.
Use
$P/BUILDENV.linux.OPENSSL_LINKAGE-master := static
$P/DEP-COMPS = $($P/DEP-COMPS) openssl/enable/static
for static compilation.
Add a large error message when both or neither are set

* Additional akamake changes

** Update crypto/evp/Makefile dependencies
** Declare our deps properly in the shared case
** Enable bf and cast for shared library (postgresql)
** Enable MD4, RipeMD, RC4, EGD to available ciphers
** Enable chacha, poly1305
** Move some variables from VAR to $P/VAR
** Add more explicit rules
** Require akamai_lib 6.5 if present

Original Authors: bkaduk@akamai.com, tshort@akamai.com
Code Reviewers: pbohman@akamai.com, bkaduk@akamai.com, tshort@akamai.com
Akamai Status: internal only