CR 3374441: Update Configuration/akamake for shared
Combination of most of the akamake-related changes for shared.

Further design discussions have led us to revisit the decision to
require components to select between static or shared linkage.
Enough opportunities are available to insert checks into release
processes that ensure that the needed shared library components are
present in releases with components that link against openssl shared
libraries, even if they do so unknowingly by upgrading to a version
of openssl that is shared by default.

Requiring component owners to explicitly pick one of shared of static
presents a number of complications, both in implementation logic
and in the fallout for release engineers.  It is more prudent to
just go back to defaulting to shared libraries with an opt-out.

Whether the openssl linkage is shared or static is a property of
the build environment, and must be global for all compilations within
a given tree.  In particular, we do not want a single component that
opts out of shared linkage for itself to cause a cascade of static
linkage for all components that include it -- only the master component
should be able to control whether that build is done with shared or
static openssl, which is exactly what the BUILDENV feature of Akamake
is designed for.  Unfortunately, changes to Akamake are required in
order to add new variables to the BUILDENV framework, but it is possible
to get a close approximation without Akamake support, at the cost of
more logic in the openssl makefile.  We can use that logic until our
patches to add support to Akamake for the OPENSSL_LINKAGE BUILDENV
variable are deployed.

In order to do as much as possible on the first (*) pass through akamake,
conditionalize the presented build targets and exported libraries based on
either (1) BUILDENV.OPENSSL_LINKAGE-master, if defined by Akamake, or
(2) whether the master component DEP-COMPS on openssl/enable/static.
We will document that components must set both, and once Akamake has support
for the OPENSSL_LINKAGE buildenv, that will take precedence.  Since we want
the master component to have control over the static opt-out, it is sufficient
to have our build logic conditionalize on the master component's DEP-COMPS.
We will also need to have a check and error out if openssl/enable/static
is present but *not* set by the master component, but that can be done
elsewhere (and must use delayed evaulation).

Provide assertions for those conditions that we expect to only
hold simultaneously.

(*) first real execution, i.e., after akamaster has processed buildenv and
akamake-env.mak and such.

* Split CONFIGFLAGS into features and flags

We will want to use a specific set of features for shared-library builds,
so use a separate varaible to facilitate reuse of the logic for the other
arguments to Configure, including the sysname and certain CFLAGS.

* Build a shared library by default

Force a fixed featureset for the shared library build. The fixed
featureset differs between 32- and 64-bit builds, because the
ec_nistp_64_gcc_128 code relies on 128-bit integers, which are only
available in the 64-bit target.

Detect the presence of openssl/enable/feature components for features not
enabled in the shared build, and error out the build with a message
noting that a static build is required for those features.

Only export the shared library for shared-library builds, to prevent
consumers from inadvertently linking openssl statically.

This does require some extremely close coordination between
openssl/akamake and openssl/(en|dis)able/*/akamake, since we must
ensure correct behavior regardless of which order they are processed in.

* Export shared libs as a tree

We already export them as EXP-LIBS, but we also need to export them
as an EXP-TREE, so that the openssl-shared component(s) can import
them and put them in a bundle.

The unversioned path is just a symlink, so it is better to export
the actual files.

* Add LD_LIBRARY_PATH for tests

Appending to LD_LIBRARY_PATH in the simple way leaves an empty
component if LD_LIBRARY_PATH was unset prior to the append operation;
this causes the current directory to be searched first, which
is usually not desired.

* Update componentinfo.xml to the 7.x version

* Additional akamake changes

** Update crypto/evp/Makefile dependencies
** Declare our deps properly in the shared case
** Enable bf and cast for shared library (postgresql)
** Enable MD4, RipeMD, RC4, EGD to available ciphers
** Enable chacha, poly1305
** Move some variables from VAR to $P/VAR
** Add more explicit rules
** Require akamai_lib 6.5 if present

Original Authors: bkaduk@akamai.com, tshort@akamai.com
Code Reviewers: pbohman@akamai.com, bkaduk@akamai.com, tshort@akamai.com
Akamai Status: internal only