CR 3452531: provide uint64_t for windows builds
The EVP_CHACHA_AEAD_CTX structure uses the uint64_t type, which is
not available on windows with the set of headers included by that
file prior to this commit.  Starting with Visual Studio 2010, the
type definition is avialable in <stdint.h>, but in previous versions
it seems that the __int64 type is a builtin (but not __uint64), allowing
a manual typedef to be used for those systems.

Original Author: bkaduk@akamai.com
Code Reviewer: pbohman@akamai.com
Akamai Status: internal-only (master addresses the issue in e_os2.h)