CR 3501161: Session resume not working
When an SSL's context is swtiched from a ticket-enabled context to a
ticket-disabled context, no session-id is generated, so the session
can't be resumed.

If a servername callback changes the SSL_OP_NO_TICKET option, check
to see if it's changed to disable, and whether a session ticket is
expected (i.e. the client indicated ticket support and the SSL had
tickets enabled at the time), and whether we already have a previous
session (i.e. s->hit is set).

In this case, clear the ticket-expected flag and create a new session
that has a session-id associated with it.

If the SSL hit (resumed) and switched to a ticket-disabled context,
assume that the resumption was via session-id, and don't bother to
create a new session.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: PR1529