CR MANY: Session resume broken switching contexts
Includes
CR 3501161: Session resume broken switching contexts
CR 3664193: Fix session id resume issue
CR 3691142: fix --strict-warnings build with gcc

CR 3501161: Session resume broken switching contexts

When an SSL's context is swtiched from a ticket-enabled context to
a ticket-disabled context in the servername callback, no session-id
is generated, so the session can't be resumed.

If a servername callback changes the SSL_OP_NO_TICKET option, check
to see if it's changed to disable, and whether a session ticket is
expected (i.e. the client indicated ticket support and the SSL had
tickets enabled at the time), and whether we already have a previous
session (i.e. s->hit is set).

In this case, clear the ticket-expected flag and create a new session
that has a session-id associated with it.

If the SSL hit (resumed) and switched to a ticket-disabled context,
assume that the resumption was via session-id, and don't bother to
create a new session.

Before this fix, the updated unit-tests in 06-sni-ticket.conf would
fail test #4 (server1 = SNI, server2 = no SNI).

CR 3691142: fix --strict-warnings build with gcc

clang-3.6 doesn't complain about our code, but the gcc-4.8.4-2ubuntu1~14.04.3
that comes with the current Akamai Ubuntu 14.0.4.5 takes issue with
trailing commas in enumeration declarations, as well as comparing integers
with enum values.

CR 3664193: Fix session id resume issue

When tickets are disabled by the SNI, a new session was created which
was discarding data. Instead, discard only the ticket data, and
generate a new session ID. This was originally CR3501161 (and can be
squashed with that).

Updated errors, which re-ordered some things.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: PR1529