CR 3374441: improve non-linux support
With all the excitement to build shared libraries, and use them
when openssl is the master component, we lost track of the part
where we only support shared libraries on linux.  So, bring back
the windows (and OS X?) support.

So, only set BUILDENV.OPENSSL_LINKAGE-master to shared when building on linux
(and for consistency, set it to static otherwise).  In the non-linux case,
we also have to set $P/NOSHARED to get the right logic later on in the
akamake file.

Also, go back to using separate definitions of CONFIGURE_STEP for windows
and non-windows.  While we could write a CONFIGURE_SHARED step for windows
that would never get used, it doesn't really represent what's going on,
which is that windows still only has one CONFIGURE_STEP that we might use.

To be squashed into commit 29964d7058b5230bfe49ea54fac42a4d1d1b6455 on rebase.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal-only